# Conflicts:
#	CMakeLists.txt
#	Dockerfile
#	deps/gtest
#	src/main/cpp/controller/controller.cc
#	src/main/cpp/controller/controller.h
#	src/main/cpp/controller/patch_builder.cc
#	src/main/cpp/controller/patch_builder_streaming.h
#	src/main/cpp/controller/snapshot_creation_strategy.h
#	src/main/cpp/controller/snapshot_patch_iterator_triple_id.cc
#	src/main/cpp/controller/snapshot_patch_iterator_triple_id.h
#	src/main/cpp/controller/triple_delta_iterator.cc
#	src/main/cpp/controller/triple_delta_iterator.h
#	src/main/cpp/controller/triple_versions_iterator.cc
#	src/main/cpp/controller/triple_versions_iterator.h
#	src/main/cpp/dictionary/dictionary_manager.cc
#	src/main/cpp/evaluate/evaluator.cc
#	src/main/cpp/evaluate/evaluator.h
#	src/main/cpp/insert.cc
#	src/main/cpp/patch/patch.h
#	src/main/cpp/patch/patch_element_iterator.cc
#	src/main/cpp/patch/patch_element_iterator.h
#	src/main/cpp/patch/patch_tree.cc
#	src/main/cpp/patch/patch_tree_addition_value.cc
#	src/main/cpp/patch/patch_tree_iterator.cc
#	src/main/cpp/patch/patch_tree_key_comparator.cc
#	src/main/cpp/patch/patch_tree_manager.cc
#	src/main/cpp/patch/patch_tree_manager.h
#	src/main/cpp/patch/triple.cc
#	src/main/cpp/patch/triple.h
#	src/main/cpp/patch/triple_comparator.cc
#	src/main/cpp/patch/triple_comparator.h
#	src/main/cpp/patch/triple_store.cc
#	src/main/cpp/patch/triple_store.h
#	src/main/cpp/query_delta_materialized.cc
#	src/main/cpp/query_version.cc
#	src/main/cpp/snapshot/combined_triple_iterator.h
#	src/main/cpp/snapshot/snapshot_manager.cc
#	src/main/cpp/snapshot/snapshot_manager.h
#	src/test/cpp/bear.cc
#	src/test/cpp/controller/controller.cc
#	src/test/cpp/dictionary/dictionary_manager.cc
#	src/test/cpp/patch/patch_tree.cc
#	src/test/cpp/patch/patch_tree_key_comparator.cc
#	src/test/cpp/patch/triple.cc
#	src/test/cpp/snapshot/snapshot_manager.cc

+++ src/main/cpp/controller/snapshot_creation_strategy.h
@@ -6,8 +6,8 @@
 
 struct CreationStrategyMetadata {
     int patch_id = 0;
-    std::vector<size_t> delta_sizes;
-    std::vector<size_t> agg_delta_sizes;
+    std::vector<uint64_t> delta_sizes;
+    std::vector<uint64_t> agg_delta_sizes;
     size_t last_snapshot_size = 0;
     std::vector<uint64_t> version_sizes;
     std::vector<uint64_t> ingestion_times;
